vscode: allow running in browser when build tool defines process.env
Many frontend build configurations define process.env even in code
running in the browser. If process is defined but process.getuid() is
not defined, then this _isRootUser check will throw an exception.

If there is no process.getuid, then the current user is likely not
root (and, in fact, the code is probably running in an environment
with no notion of a "root user", such as the browser).